new-policy-play/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
motor>=3.3.2
pydantic[email]>=2.9.0
email-validator>=2.0.0
groq>=0.4.1
pdfplumber>=0.10.3
mammoth>=1.6.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<5.0.0
certifi>=2023.7.22

